requirements.txt
davia
pandas
python-dotenv
google-generativeai
